// src/api/__generated__/mining/mining/mining.ts
export type listRestActorsResponse200 = {
  data: PageActorDTO;
  status: 200;
};

export type listRestActorsResponse400 = {
  data: ResultObject;
  status: 400;
};

export type listRestActorsResponse404 = {
  data: ResultObject;
  status: 404;
};

export type listRestActorsResponse500 = {
  data: ResultObject;
  status: 500;
};

export type listRestActorsResponseSuccess = listRestActorsResponse200 & {
  headers: Headers;
};
export type listRestActorsResponseError = (
  | listRestActorsResponse400
  | listRestActorsResponse404
  | listRestActorsResponse500
) & {
  headers: Headers;
};

export type listRestActorsResponse =
  | listRestActorsResponseSuccess
  | listRestActorsResponseError;
// ...
export type listDeployedActorsResponse200 = {
  data: ActorDTO[];
  status: 200;
};

export type listDeployedActorsResponse400 = {
  data: ResultObject;
  status: 400;
};

export type listDeployedActorsResponse404 = {
  data: ResultObject;
  status: 404;
};

export type listDeployedActorsResponse500 = {
  data: ResultObject;
  status: 500;
};

export type listDeployedActorsResponseSuccess =
  listDeployedActorsResponse200 & {
    headers: Headers;
  };
export type listDeployedActorsResponseError = (
  | listDeployedActorsResponse400
  | listDeployedActorsResponse404
  | listDeployedActorsResponse500
) & {
  headers: Headers;
};

export type listDeployedActorsResponse =
  | listDeployedActorsResponseSuccess
  | listDeployedActorsResponseError;
// ...
export type listAllActorsResponse200 = {
  data: PageActorDTO;
  status: 200;
};

export type listAllActorsResponse400 = {
  data: ResultObject;
  status: 400;
};

export type listAllActorsResponse404 = {
  data: ResultObject;
  status: 404;
};

export type listAllActorsResponse500 = {
  data: ResultObject;
  status: 500;
};

export type listAllActorsResponseSuccess = listAllActorsResponse200 & {
  headers: Headers;
};
export type listAllActorsResponseError = (
  | listAllActorsResponse400
  | listAllActorsResponse404
  | listAllActorsResponse500
) & {
  headers: Headers;
};

export type listAllActorsResponse =
  | listAllActorsResponseSuccess
  | listAllActorsResponseError;