// src/api/__generated__/mining/model/pageActorDTO.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Mining API
 * 挖矿服务接口
 * OpenAPI spec version: v1.0
 */
import type { ActorDTO } from './actorDTO';

export interface PageActorDTO {
  records?: ActorDTO[];
  pageNumber?: number;
  pageSize?: number;
  totalPage?: number;
  totalRow?: number;
  optimizeCountQuery?: boolean;
}